Exception handling

+++ dscommerce/src/main/java/com/jpgedvila/dscommerce/services/ProductService.java
@@ -3,9 +3,14 @@ package com.jpgedvila.dscommerce.services;
 import com.jpgedvila.dscommerce.dto.ProductDTO;
 import com.jpgedvila.dscommerce.entities.Product;
 import com.jpgedvila.dscommerce.repositories.ProductRepository;
+import com.jpgedvila.dscommerce.services.exceptions.DatabaseException;
+import com.jpgedvila.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -19,7 +24,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -43,17 +49,29 @@ public class ProductService {
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
 
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
+        try{
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
 
-        entity = repository.save(entity);
+            entity = repository.save(entity);
 
-        return new ProductDTO(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e){
+            throw new  ResourceNotFoundException("Recurso não encontrado");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try{
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referêncial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
